Split ISC\Image_Sources\Admin
Split the class into multiple smaller ones and mainly use the existing one to load the other components.
