+++ includes/image-sources/admin/scripts.php
@@ -1,0 +1,80 @@
+<?php
+
+namespace ISC\Image_Sources;
+
+use ISC\Admin_Utils;
+
+/**
+ * Add the admin menu items für Image Sources features
+ */
+class Image_Sources_Admin_Scripts {
+	/**
+	 * Constructor
+	 */
+	public function __construct() {
+		add_action( 'admin_enqueue_scripts', [ $this, 'add_admin_scripts' ] );
+		add_action( 'admin_print_scripts', [ $this, 'admin_head_scripts' ] );
+	}
+
+	/**
+	 * Add scripts to ISC-related pages
+	 */
+	public function add_admin_scripts() {
+		$screen = get_current_screen();
+
+		if ( ! isset( $screen->id ) ) {
+			return;
+		}
+
+		if ( $screen->id === 'media_page_isc-sources' ) {
+			wp_enqueue_script( 'isc_sources_script', ISCBASEURL . '/admin/assets/js/sources.js', [], ISCVERSION, true );
+		}
+
+		if ( in_array( $screen->id, [ 'upload', 'widgets', 'customize' ], true ) ) {
+			wp_enqueue_script( 'isc_attachment_compat', ISCBASEURL . '/admin/assets/js/wp.media.view.AttachmentCompat.js', [ 'media-upload' ], ISCVERSION, true );
+		}
+	}
+
+	/**
+	 * Display scripts in <head></head> section of admin page. Useful for creating js variables in the js global namespace.
+	 */
+	public function admin_head_scripts() {
+		global $pagenow;
+		$screen = get_current_screen();
+		// texts in JavaScript on sources page
+		if ( 'upload.php' === $pagenow && isset( $_GET['page'] ) && 'isc-sources' === $_GET['page'] ) {
+			?>
+			<script>
+				isc_data = {
+					confirm_message: '<?php esc_html_e( 'Are you sure?', 'image-source-control-isc' ); ?>',
+					baseurl:         '<?php echo esc_url( ISCBASEURL ); ?>'
+				};
+			</script>
+			<?php
+		}
+		// add style to media edit pages
+		if ( isset( $screen->id ) && $screen->id === 'attachment' ) {
+			// Meta field in media view
+			?>
+			<style>
+				.compat-attachment-fields input[type="text"] {
+					width: 100%;
+				}
+
+				.compat-attachment-fields th {
+					vertical-align: top;
+				}
+			</style>
+			<?php
+		}
+		// add to any backend pages
+		?>
+		<style>
+			.compat-attachment-fields .isc-get-pro {
+				font-weight: bold;
+				color: #F70;
+			}
+		</style>
+		<?php
+	}
+}
