+++ lib/composer/autoload_classmap.php
@@ -8,13 +8,19 @@ $baseDir = dirname($vendorDir);
 return array(
     'Composer\\InstalledVersions' => $vendorDir . '/composer/InstalledVersions.php',
     'ISC\\Admin' => $baseDir . '/admin/admin.php',
+    'ISC\\Admin\\Admin_Scripts' => $baseDir . '/admin/includes/scripts.php',
     'ISC\\Admin_Utils' => $baseDir . '/admin/utils.php',
     'ISC\\Autoloader' => $baseDir . '/includes/class-autoloader.php',
     'ISC\\Feedback' => $baseDir . '/includes/feedback.php',
     'ISC\\Helpers' => $baseDir . '/includes/helpers.php',
     'ISC\\Image_Sources\\Admin' => $baseDir . '/includes/image-sources/admin.php',
+    'ISC\\Image_Sources\\Admin_Ajax' => $baseDir . '/includes/image-sources/admin/ajax.php',
+    'ISC\\Image_Sources\\Admin_Fields' => $baseDir . '/includes/image-sources/admin/fields.php',
+    'ISC\\Image_Sources\\Admin_Menu' => $baseDir . '/includes/image-sources/admin/menu.php',
+    'ISC\\Image_Sources\\Admin_Notices' => $baseDir . '/includes/image-sources/admin/notices.php',
     'ISC\\Image_Sources\\Analyze_HTML' => $baseDir . '/includes/image-sources/analyze-html.php',
     'ISC\\Image_Sources\\Image_Sources' => $baseDir . '/includes/image-sources/image-sources.php',
+    'ISC\\Image_Sources\\Image_Sources_Admin_Scripts' => $baseDir . '/includes/image-sources/admin/scripts.php',
     'ISC\\Image_Sources\\Utils' => $baseDir . '/includes/image-sources/utils.php',
     'ISC\\Indexer' => $baseDir . '/includes/indexer.php',
     'ISC\\Newsletter' => $baseDir . '/includes/newsletter.php',
