+++ lib/composer/autoload_static.php
@@ -28,13 +28,19 @@ class ComposerStaticInitdb72a6bac11cb0e6b971811c93d09a9b
     public static $classMap = array (
         'Composer\\InstalledVersions' => __DIR__ . '/..' . '/composer/InstalledVersions.php',
         'ISC\\Admin' => __DIR__ . '/../..' . '/admin/admin.php',
+        'ISC\\Admin\\Admin_Scripts' => __DIR__ . '/../..' . '/admin/includes/scripts.php',
         'ISC\\Admin_Utils' => __DIR__ . '/../..' . '/admin/utils.php',
         'ISC\\Autoloader' => __DIR__ . '/../..' . '/includes/class-autoloader.php',
         'ISC\\Feedback' => __DIR__ . '/../..' . '/includes/feedback.php',
         'ISC\\Helpers' => __DIR__ . '/../..' . '/includes/helpers.php',
         'ISC\\Image_Sources\\Admin' => __DIR__ . '/../..' . '/includes/image-sources/admin.php',
+        'ISC\\Image_Sources\\Admin_Ajax' => __DIR__ . '/../..' . '/includes/image-sources/admin/ajax.php',
+        'ISC\\Image_Sources\\Admin_Fields' => __DIR__ . '/../..' . '/includes/image-sources/admin/fields.php',
+        'ISC\\Image_Sources\\Admin_Menu' => __DIR__ . '/../..' . '/includes/image-sources/admin/menu.php',
+        'ISC\\Image_Sources\\Admin_Notices' => __DIR__ . '/../..' . '/includes/image-sources/admin/notices.php',
         'ISC\\Image_Sources\\Analyze_HTML' => __DIR__ . '/../..' . '/includes/image-sources/analyze-html.php',
         'ISC\\Image_Sources\\Image_Sources' => __DIR__ . '/../..' . '/includes/image-sources/image-sources.php',
+        'ISC\\Image_Sources\\Image_Sources_Admin_Scripts' => __DIR__ . '/../..' . '/includes/image-sources/admin/scripts.php',
         'ISC\\Image_Sources\\Utils' => __DIR__ . '/../..' . '/includes/image-sources/utils.php',
         'ISC\\Indexer' => __DIR__ . '/../..' . '/includes/indexer.php',
         'ISC\\Newsletter' => __DIR__ . '/../..' . '/includes/newsletter.php',
